Add files via upload
To obtain a simple formula the problem must be idealised somewhat, so the egg will be treated as a spherical homogeneous object of mass M and initial temperature Tegg. If the egg is placed straight into a pan of boiling water at Twater, it will be ready when the temperature at the boundary of the yolk has risen to Tyolk~63°C. With these assumptions, the cooking time t can be deduced by solving a heat diffusion equation.